# server/swagger.yaml
openapi: 3.0.3
info:
  title: Milk Dairy Management System API
  description: |
    A comprehensive MERN stack application for managing milk dairy operations with role-based access control, 
    subscription management, and real-time delivery tracking.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## User Roles
    - **ADMIN**: Full system access, user management, analytics
    - **DISTRIBUTOR**: Order management, delivery assignment, product management
    - **DELIVERY_BOY**: Delivery assignments, status updates
    - **CONSUMER**: Place orders, manage subscriptions, track deliveries
  version: 1.0.0
  contact:
    name: API Support
    email: support@milkdairy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.milkdairy.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /addresses:
    get:
      tags:
        - Address Management
      summary: Get all user addresses
      description: Retrieve all addresses for the authenticated user
      responses:
        '200':
          description: Addresses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Address Management
      summary: Add new address
      description: Create a new address for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
            example:
              name: "John Doe"
              phone: "9876543210"
              street: "123 Main Street, Apartment 4B"
              city: "New Delhi"
              state: "Delhi"
              pincode: "110001"
              landmark: "Near Metro Station"
              isDefault: true
      responses:
        '201':
          description: Address added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Address'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /subscriptions:
    get:
      tags:
        - Subscription Management
      summary: Get user subscriptions
      description: Retrieve all subscriptions for the authenticated user
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Subscription Management
      summary: Create new subscription
      description: Create a new subscription for recurring orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
            example:
              products:
                - productId: "60d5ecb74b24a1234567890a"
                  quantity: 2
              subscriptionType: "weekly"
              frequency: "weekly"
              startDate: "2024-01-01"
              endDate: "2024-12-31"
              deliveryTime: "morning"
              deliveryDays: ["monday", "wednesday", "friday"]
              specialInstructions: "Leave at the door"
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subscriptions/{id}:
    get:
      tags:
        - Subscription Management
      summary: Get subscription by ID
      description: Retrieve a specific subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Subscription Management
      summary: Update subscription
      description: Update an existing subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Subscription Management
      summary: Cancel subscription
      description: Cancel an existing subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Subscription already cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /consumer/dashboard:
    get:
      tags:
        - Consumer
      summary: Get consumer dashboard data
      description: Retrieve dashboard data including recent orders and deliveries
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
                      deliveries:
                        type: array
                        items:
                          $ref: '#/components/schemas/Delivery'
                      totalOrders:
                        type: integer
                      totalDeliveries:
                        type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consumer/orders:
    get:
      tags:
        - Consumer
      summary: Get consumer orders
      description: Retrieve all orders for the authenticated consumer
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Consumer
      summary: Place new order
      description: Create a new order for products from assigned distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
            example:
              products:
                - productId: "60d5ecb74b24a1234567890a"
                  quantity: 2
              deliveryDate: "2024-01-15"
              deliveryTime: "morning"
              deliveryAddress:
                street: "123 Main Street"
                city: "New Delhi"
                state: "Delhi"
                pincode: "110001"
                landmark: "Near Metro Station"
              specialInstructions: "Ring the bell twice"
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or no distributor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consumer/products:
    get:
      tags:
        - Consumer
      summary: Get available products
      description: Get products available from the assigned distributor
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '400':
          description: No distributor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Consumer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consumer/subscriptions:
    get:
      tags:
        - Consumer
      summary: Get consumer subscriptions
      description: Retrieve all subscriptions for the authenticated consumer
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Consumer
      summary: Create consumer subscription
      description: Create a new subscription for recurring orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerSubscriptionRequest'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or no distributor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consumer/subscriptions/{subscriptionId}:
    put:
      tags:
        - Consumer
      summary: Update consumer subscription
      description: Update an existing subscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '403':
          description: Not authorized to update this subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        _id:
          type: string
          description: Address ID
        userId:
          type: string
          description: User ID who owns this address
        name:
          type: string
          description: Name for this address
        phone:
          type: string
          pattern: '^[0-9]{10}$'
          description: Phone number for this address
        street:
          type: string
          description: Street address
        city:
          type: string
          description: City name
        state:
          type: string
          description: State name
        pincode:
          type: string
          pattern: '^[0-9]{6}$'
          description: 6-digit pincode
        landmark:
          type: string
          description: Nearby landmark (optional)
        isDefault:
          type: boolean
          description: Whether this is the default address
        coordinates:
          type: object
          properties:
            lat:
              type: number
              minimum: -90
              maximum: 90
              description: Latitude
            lng:
              type: number
              minimum: -180
              maximum: 180
              description: Longitude
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (development only)
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Validation errors
...